+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## [2.0.2] - 2017-03.12
+
+### Changed
+Improved FTP support for non-compliant servers
+Fix for ZipFS implied directories
+
 ## [2.0.1] - 2017-03-11
 
 ### Added